+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## [dev]
+- deprecate `SignalsAPI.submission_status`
+
 ## [2.13.4] - 2023-04-03
 - return filepath from `download_dataset`